+++ Interfaces/GCHP/Chem_GridCompMod.F90
@@ -2703,12 +2703,31 @@ CONTAINS
           ! Get Internal state
           CALL MAPL_Get ( STATE, INTERNAL_ESMF_STATE=INTERNAL, __RC__ )
 
+
+#ifdef JACOBIAN
+          ! Set Jacobian local variables
+          primarySpcName = ''
+          primarySpcId = -1
+          DO N = 1, State_Chm%nSpecies
+             ThisSpc => State_Chm%SpcData(N)%Info
+             IF ( ThisSpc%Is_JacobianTracer .AND. primarySpcID == -1 ) THEN
+                ThisSpc => State_Chm%SpcData(N)%Info
+                primarySpcName = ThisSpc%Name(1:LEN(trim(ThisSpc%Name))-8)
+                primarySpcId = IND_(trim(primarySpcName))
+                EXIT
+             ENDIF
+          ENDDO
+#endif
+
           ! Loop over all species and get info from spc db
           DO N = 1, State_Chm%nSpecies
              ThisSpc => State_Chm%SpcData(N)%Info
              IF ( TRIM(ThisSpc%Name) == '' ) CYCLE
              IND = IND_( TRIM(ThisSpc%Name ) )
              IF ( IND < 0 ) CYCLE
+#ifdef JACOBIAN
+             IF ( ThisSpc%Is_JacobianTracer ) CYCLE
+#endif
 
              ! Determine if species in restart file
              CALL ESMF_StateGet( INTERNAL, TRIM(SPFX) // TRIM(ThisSpc%Name),  &
@@ -2717,13 +2736,20 @@ CONTAINS
                   VALUE=RST, RC=STATUS )
 
              ! Set spc conc to background value if rst skipped or var not there
-             IF ( ( RC  /= ESMF_SUCCESS           .OR.     &
-                    RST == MAPL_RestartBootstrap  .OR.     &
-                    RST == MAPL_RestartSkipInitial  )      &
+             IF ( ( RC  /= ESMF_SUCCESS         .OR.     &
+                  RST == MAPL_RestartBootstrap  .OR.     &
+                  RST == MAPL_RestartSkipInitial  ) THEN
+
 #ifdef JACOBIAN
-                    .AND. .NOT. ThisSpc%Is_JacobianTracer  &
+                ! Exit with error if the primary species is not found in restart
+                IF ( primarySpcId == IND ) THEN
+                   WRITE(*,*) '   ERROR: Cannot find primary species used to initialize ' &
+                        'Jacobian tracers in restart file: ' // trim(primarySpcName)
+                   STATUS = GC_FAILURE
+                   _VERIFY(STATUS)
+                ENDIF
 #endif
-                    ) THEN
+
                 DO L = 1, State_Grid%NZ
                 DO J = 1, State_Grid%NY
                 DO I = 1, State_Grid%NX
@@ -2745,11 +2771,22 @@ CONTAINS
                 ENDIF
              ENDIF
 
+             ThisSpc => NULL()
+          ENDDO
+
 #ifdef JACOBIAN
-             ! Do special handling if this is a Jacobian tracer             
-             IF ( ThisSpc%Is_JacobianTracer ) THEN
-                primarySpcName = ThisSpc%Name(1:LEN(trim(ThisSpc%Name))-8)
-                primarySpcId = IND_(trim(primarySpcName))
+          ! Set Jacobian species to primary species values if the Jacobian
+          ! species is missing in the restart file
+          DO N = 1, State_Chm%nSpecies
+             ThisSpc => State_Chm%SpcData(N)%Info
+             IF ( .NOT. ThisSpc%Is_JacobianTracer ) CYCLE
+
+             CALL ESMF_StateGet( INTERNAL, TRIM(SPFX) // TRIM(ThisSpc%Name),  &
+                  trcFIELD, RC=RC )
+             CALL ESMF_AttributeGet( trcFIELD, NAME="RESTART", VALUE=RST, RC=STATUS )
+             IF ( RC  /= ESMF_SUCCESS           .OR.     &
+                  RST == MAPL_RestartBootstrap  .OR.     &
+                  RST == MAPL_RestartSkipInitial  ) THEN
                 State_Chm%Species(IND)%Conc = State_Chm%Species(primarySpcId)%Conc
                 IF ( MAPL_am_I_Root()) THEN
                    WRITE(*,*) '   INFO: using the initial concentration of ' &
@@ -2757,10 +2794,11 @@ CONTAINS
                         // trim(ThisSpc%Name)
                 ENDIF
              ENDIF
-#endif
 
              ThisSpc => NULL()
           ENDDO
+#endif
+
        ENDIF
 
        !=======================================================================